Reorganize conf
