+++ commons/zenoh-config/src/qos.rs
@@ -15,6 +15,8 @@ use serde::{Deserialize, Serialize};
 use zenoh_keyexpr::keyexpr_tree::{IKeyExprTreeMut, KeBoxTree};
 use zenoh_protocol::core::{key_expr::OwnedKeyExpr, CongestionControl, Priority, Reliability};
 
+use crate::InterceptorFlow;
+
 #[derive(Debug, Deserialize, Default, Serialize, Clone)]
 pub struct PublisherQoSConfList(pub(crate) Vec<PublisherQoSConf>);
 
@@ -130,6 +132,32 @@ pub enum PublisherLocalityConf {
     SessionLocal,
     Remote,
     Any,
+}
+
+#[derive(Default, Debug, Deserialize, Serialize, Clone)]
+pub struct QosOverwriteItemConf {
+    /// Optional identifier for the qos modification configuration item.
+    pub id: Option<String>,
+    /// A list of interfaces to which the qos will be applied.
+    /// QosOverwrite will be applied for all interfaces if the parameter is None.
+    pub interfaces: Option<Vec<String>>,
+    /// List of message types on which the qos overwrite will be applied.
+    pub messages: Vec<QosOverwriteMessage>,
+    /// List of key expressions to apply qos overwrite.
+    pub key_exprs: Vec<OwnedKeyExpr>,
+    // The qos value to overwrite with.
+    pub overwrite: QosOverwrites,
+    /// QosOverwrite flow directions: egress and/or ingress.
+    pub flows: Option<Vec<InterceptorFlow>>,
+}
+
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, Eq, Hash, PartialEq)]
+#[serde(rename_all = "snake_case")]
+pub enum QosOverwriteMessage {
+    Put,
+    Delete,
+    Query,
+    Reply,
 }
 
 #[derive(Debug, Default, Deserialize, Serialize, Clone)]
